Blender Bezier Parser
Was finally able to get the blender tool ready. Still needs a bit of polishing, but overall it works. This allowed me to make the bezier surface a bit more complex, allowing for hard edges. The mathematics is now more accurate to the definition of the surface. Before we interpolated curves between curves. Now we set up an array of points and interpolate between those. Despite the complexity, it is way simpler to compute that the first implementation.