Fixed Parser and working on tutorial Doc
The parser works as intended now. Will move it to the Canoe Example directory later. The tutorial doc is being reworked for the haida canoe since there are more complex shapes there and allow me to show how complex you can make the canoes.

Reworked all the code on the bezier surface so that it would properly display and truely go from 0-1 on each tile.
Updated it so that the Bouyancy now accepted the new canoes.
Updated so we no longer effect the width but instead effec the scaling of the canoe. This fixes the issue where the canoes had different lengths when switching between models. Now its unified. Length width and height get displayed on the graph.