Math writing and hyperlinks to graphs
Added more math to the mathematical Canoes to help bring it up to presentation level. Added a bit more to the Designer Canoes for modelling. Worked on making desmos graphs for images, and so people could interact with the math, though not in the notebook. We'll leave that for the canoes or other simpler graphs, until we can embed desmos pages into the notebook which may or may not happen.